server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
moviepy==2.1.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.1
openai==1.40.6
numpy==2.1.0
